// src/models/book-period-soon.d.ts
export interface BookPeriodSoonModel {
    id: number;
    amount: string;
    interest: string;
    early_interest: string;
    date_pay: string;
    date_diff: string;
    tax: string;
    profit_before_tax: string;
    money_received: string;
    total_money: string;
    ma_so: string;
    so_tien_goc: string;
    ngay_bat_dau: string;
    ngay_rut_dung_han: string;
    muc_loi_nhuan_dung_han: string;
    loi_nhuan_nhan_duoc_dung_han: string;
    thue_thu_nhap_ca_nhan_dung_han: string;
    so_tien_nhan_duoc_sau_thue_dung_han: string;
    ngay_nhan_tien: string;
    lai_suat_rut_truoc_han: string;
    muc_loi_nhuan_truoc_han: string;
    thue_thu_nhap_ca_nhan_truoc_han: string;
    so_tien_nhan_duoc_sau_thue_truoc_han: string;
    loi_nhuan: string;
    loi_nhuan_bi_mat: string;
}